Skip installing if no packages to publish

+++ synchronize-with-npm/src/publishAndTag.ts
@@ -19,32 +19,36 @@ export function* publishAndTag({
   octokit,
   payload,
 }: Publish): Operation<ToPublish[]> {
-  let installCommand = installScript || fs.existsSync("yarn.lock") ? "yarn install --frozen-lockfile" : "npm ci";
-
-  console.log(
-    colors.yellow("Installing with command"),
-    colors.blue(installCommand)+colors.yellow("...\n"),
-  );
-  yield exec(installCommand).join();
-
-  let successfullyPublished: ToPublish[] = [];
-  yield all(
-    confirmedPkgsToPublish.map(pkg =>
-      function* () {
-        let result: ProcessResult = yield exec("npm publish --access=public", { cwd: pkg.path }).join();
-        // TODO how can i turn octokit.request into an operation so i can .join()
-        if (result.code === 0) {
-          yield octokit.request("POST /repos/{owner}/{repo}/git/refs", {
-            owner: payload.repository.owner.login,
-            repo: payload.repository.name,
-            ref: `refs/tags/${pkg.name}-v${pkg.version}`,
-            sha: payload.after,
-          });
-          successfullyPublished = [...successfullyPublished, pkg];
+  if (confirmedPkgsToPublish.length) {
+    let installCommand = installScript || fs.existsSync("yarn.lock") ? "yarn install --frozen-lockfile" : "npm ci";
+  
+    console.log(
+      colors.yellow("Installing with command"),
+      colors.blue(installCommand)+colors.yellow("...\n"),
+    );
+    yield exec(installCommand).join();
+  
+    let successfullyPublished: ToPublish[] = [];
+    yield all(
+      confirmedPkgsToPublish.map(pkg =>
+        function* () {
+          let result: ProcessResult = yield exec("npm publish --access=public", { cwd: pkg.path }).join();
+          // TODO how can i turn octokit.request into an operation so i can .join()
+          if (result.code === 0) {
+            yield octokit.request("POST /repos/{owner}/{repo}/git/refs", {
+              owner: payload.repository.owner.login,
+              repo: payload.repository.name,
+              ref: `refs/tags/${pkg.name}-v${pkg.version}`,
+              sha: payload.after,
+            });
+            successfullyPublished = [...successfullyPublished, pkg];
+          }
         }
-      }
-    )
-  );
-
-  return successfullyPublished;
+      )
+    );
+  
+    return successfullyPublished;
+  } else {
+    return [] as ToPublish[];
+  }
 }
